Show restart dialog when restarting from config editor
